add file: some-program-file
This just gives me a dedicated place from which to reference the SPDX headers
I typically use.

Note that there is also a GitHub Gist with this same content:

    https://gist.github.com/salewski/44d7976d3d3551481323c6798f22c41f

The current repo is more handy for my day to day work, but having the Gist
allows the first five lines of content from the Gist to be visible on my
GitHub "Overview" page, in one of my "Pinned" repo boxes.

Although the SPDX header lines appear truncated on the screen, the full
content is present, and you can copy the text directly from the "Pinned" box.