add file: some-program-file-one-line
This is an exact copy of the file:

    'some-program-file'

named to reflect the use of the single-line variant of the
'SPDX-FileCopyrightText' tag. I'll be changing the original file
shortly to use a multi-line variant that I've come to prefer for
most use cases, but want to keep the single-line version easily
accessible, as well (since it is still preferred in those places
where vertical space economy is paramount).