some-program-file, README.md: use multi-line SPDX copyright
Use the multi-line variation of the 'SPDX-FileCopyrightText' tag
that I've come to prefer for most use cases. Some benefits of this
format include:

    * Fits without wrapping in an 80 column terminal (yes, that's
      still important)

    * More easily parsed by humans

    * Allows listing of multiple copyright holders (one per line),
      where such a need exists.

The old single-line variation has been moved to the file:

    'some-program-file-one-line'

so it is still easily accessible, as well. That variation is still
preferred in those places where vertical space economy is paramount.